refacto(pallet_starknet): move more logic out of lib.rs into separated files
